migrating away from python cgi module
This module was deprecated and will be removed in 3.13

Mostly a straightforward change with one exception: the use
of cgi.FieldStorage.headers. As I understand it, those headers are only
set when handling POST requests. As htmldiff only handles GET,

I removed the call to fields and adjusted one function to read
the value of the If-Modified-Since directly from the environment.